Rename jwt-service.js -> jwt.service.js
