+++ admin/spec/routes/completed-check.spec.js
@@ -3,7 +3,7 @@
 const proxyquire = require('proxyquire').noCallThru()
 const httpMocks = require('node-mocks-http')
 const sinon = require('sinon')
-const jwtService = require('../../services/jwt-service')
+const jwtService = require('../../services/jwt.service')
 const {audit, inputs, answers} = require('../mocks/check-complete')
 
 require('sinon-mongoose')
